Generate regexp directly during build not boot
