+++ lib/marcel.rb
@@ -2,7 +2,6 @@
 
 module Marcel
   require "marcel/version"
-  require "marcel/tika_regex"
   require "marcel/magic"
   require "marcel/mime_type"
 end
